doing thickness transfer

+++ source/calculation/fitting/fitting.cpp
@@ -587,7 +587,7 @@ void Fitting::add_Fit_To_File(const gsl_vector* x, double residual, QString file
 		index++;
 
 		//------------------------------------------------------
-		for(int param_Index=0; param_Index<params.p; ++param_Index)
+		for(size_t param_Index=0; param_Index<params.p; ++param_Index)
 		{
 			double unparametrized = params.main_Calculation_Module->unparametrize(	gsl_vector_get(x, param_Index),
 																					params.fitables.param_Pointers[param_Index]->fit.min,
